changed ftplen from float to double

+++ constants.h
@@ -12,7 +12,7 @@
 #include <map>
 
 typedef std::vector<std::pair<std::string, std::vector<double> > > dataframe;
-typedef std::vector<std::pair<int, float> > ftplen;
+typedef std::vector<std::pair<int, double> > ftplen;
 
 const int COL_START = 2;	// features starting from 3rd column in dataset (starting from 0); E.g in the infile, first 2 cols are metadata i.e, ground truth and diff score
 const int NUM_REPS = 30;	// change to 30 after testing
